[MSLITE] CodeCheck and Pclint clean

+++ mindspore/lite/tools/converter/preprocess/opencv_utils.cc
@@ -68,11 +68,12 @@ int GetMatData(const cv::Mat &mat, void **data, size_t *size) {
   }
   (*size) = 0;
   for (int i = 0; i < mat.rows; ++i) {
-    (*size) += mat.cols * mat.elemSize();
+    (*size) += static_cast<size_t>(mat.cols) * mat.elemSize();
   }
 
   (*data) = new char[*size];
-  if (memcpy_s(*data, *size, mat_local.data, mat.rows * mat.cols * mat.channels() * sizeof(float)) != EOK) {
+  if (memcpy_s(*data, *size, mat_local.data,
+               static_cast<size_t>(mat.rows * mat.cols * mat.channels()) * sizeof(float)) != EOK) {
     MS_LOG(ERROR) << "memcpy failed.";
     return RET_ERROR;
   }
